% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascade_names.R
\name{cascade_names}
\alias{cascade_names}
\title{Cascade Names}
\usage{
cascade_names(list_or_df, prefix = "")
}
\description{
Cascade names takes data_frames in a nested lists and pulls
them up to the top level, appending names that it encounters
along the way
}
