% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.R
\name{col_spec}
\alias{col_spec}
\title{TEMP - Col Spec}
\usage{
col_spec(col_types, default = col_guess())
}
\arguments{
\item{col_types}{A list of column types}

\item{default}{The default col_spec}
}
\value{
A \code{col_spec} object...?
}
\description{
This object should really not be exported.  It is due to laziness
in API development, really.
}
