% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_number}
\alias{parse_number}
\title{Parse numbers, flexibly}
\usage{
parse_number(x, na = c("", "NA"), locale = default_locale(),
  trim_ws = TRUE)
}
\arguments{
\item{x}{Character vector of values to parse.}
}
\description{
This drops any non-numeric characters before or after the first number.
The grouping mark specified by the locale is ignored inside the number.
}
\examples{
parse_number("$1000")
parse_number("1,234,567.78")
}
\seealso{
Other parsers: \code{\link{parse_factor}},
  \code{\link{parse_guess}}, \code{\link{parse_logical}}
}
