% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_vector}
\alias{parse_vector}
\title{Parse a character vector.}
\usage{
parse_vector(x, collector, na = c("", "NA"), locale = default_locale(),
  trim_ws = TRUE)
}
\arguments{
\item{x}{Character vector of elements to parse.}

\item{collector}{Column specification.}
}
\description{
Parse a character vector.
}
\examples{
x <- c("1", "2", "3", "NA")
parse_vector(x, col_integer())
parse_vector(x, col_double())
}
\keyword{internal}
