% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread.R
\name{process_spec}
\alias{process_spec}
\title{Process Spec}
\usage{
process_spec(inlist, spec)
}
\arguments{
\item{inlist}{A list object}

\item{spec}{A \code{col_spec} object}
}
\value{
A tibble with \code{spec} "processed"
}
\description{
Processes a \code{spec} object against a provided list.  Does the
following:
}
\details{
\itemize{
\item Checks to see if the list has names
\itemize{
\item if so, uses the names of \code{spec$cols}
\item otherwise, tries to force names of \code{spec$cols} into numerics
}
\item Uses \code{mapply} to apply \code{process_spec_one} to the list, one \code{spec} at a time.
\item Uses \code{cascade_names} to take care of nested structures, cascade names, and bind columns together
}
}
\keyword{internal}
