% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread.R
\name{process_spec_one}
\alias{process_spec_one}
\title{Process One Spec}
\usage{
process_spec_one(inlist, col_name, collector)
}
\arguments{
\item{inlist}{A list object}

\item{col_name}{The column name that will be extracted from the list}

\item{collector}{The collector object that will be used to parse the list}
}
\value{
tibble or list of tibbles with the new columns
}
\description{
Processes a singluar "column" of the \code{spec} being processed.  It completes the following
in order:
\itemize{
\item Extract the desired item
\item Pull out the "value" object returned
\item Ensure that the class of the value has been preserved appropriately
\item Checks to see if any recursive calls are needed (\code{length(collector) > 0})
\item Calls any recursive calls necessary
\item Sets names
\item Returns a \code{tibble} if a \code{tibble} is not already present
}
}
\keyword{internal}
