% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spread.R
\name{pull_item}
\alias{pull_item}
\title{Pull Item}
\usage{
pull_item(inlist, target, collector)
}
\arguments{
\item{inlist}{A list object}

\item{target}{The targeted key to extract (a length 1 character vector or numeric vector)}

\item{collector}{A \code{collector} object to use when parsing`}
}
\value{
A list of "value" and "list".  "value" is the extracted value, "list" is
the resulting list object that should replace the original
}
\description{
Extracts \code{target} from \code{inlist}, using \code{collector} and \code{parse}.
Also checks to ensure that the incoming value is a list, and
has a start towards "modifying-on-reference" the original list item
}
\keyword{internal}
